商品の状態カテゴリ追加
