--- app/controllers/messages_controller.rb
@@ -1,4 +1,0 @@
-class MessagesController < ApplicationController
-  def index
-  end
-end
