--- app/helpers/messages_helper.rb
@@ -1,2 +1,0 @@
-module MessagesHelper
-end
