--- app/models/article.rb
@@ -1,11 +1,0 @@
-class Article < ApplicationRecord
-  extend ActiveHash::Associations::ActiveRecordExtensions
-  belongs_to :genre
-  
-  validates :genre_id, numericality: { other_than: 1, message: "can't bi blank" }
-
-
-
-
-
-end
