--- app/views/categories/index.html.erb
@@ -1,2 +1,0 @@
-<h1>Categories#index</h1>
-<p>Find me in app/views/categories/index.html.erb</p>
