--- app/views/categories/new.html.erb
@@ -1,2 +1,0 @@
-<h1>Categories#new</h1>
-<p>Find me in app/views/categories/new.html.erb</p>
