--- app/views/messages/index.html.erb
@@ -1,2 +1,0 @@
-<h1>Messages#index</h1>
-<p>Find me in app/views/messages/index.html.erb</p>
