--- db/migrate/20231022124226_create_articles.rb
@@ -1,11 +1,0 @@
-class CreateArticles < ActiveRecord::Migration[7.0]
-  def change
-    create_table :articles do |t|
-     t.string     :title        , null: false
-     t.text       :text         , null: false
-     t.integer    :genre_id     , null: false
-
-      t.timestamps
-    end
-  end
-end
