--- db/migrate/20231022125911_create_categories.rb
@@ -1,10 +1,0 @@
-class CreateCategories < ActiveRecord::Migration[7.0]
-  def change
-    create_table :categories do |t|
-      t.string     :title        , null: false
-      t.text       :text         , null: false
-      t.integer    :genre_id     , null: false
-      t.timestamps
-    end
-  end
-end
