--- spec/factories/articles.rb
@@ -1,5 +1,0 @@
-FactoryBot.define do
-  factory :article do
-    
-  end
-end
