--- spec/factories/categories.rb
@@ -1,5 +1,0 @@
-FactoryBot.define do
-  factory :category do
-    
-  end
-end
