--- spec/factories/situations.rb
@@ -1,5 +1,0 @@
-FactoryBot.define do
-  factory :situation do
-    
-  end
-end
