--- spec/requests/categories_request_spec.rb
@@ -1,19 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe "Categories", type: :request do
-
-  describe "GET /index" do
-    it "returns http success" do
-      get "/categories/index"
-      expect(response).to have_http_status(:success)
-    end
-  end
-
-  describe "GET /new" do
-    it "returns http success" do
-      get "/categories/new"
-      expect(response).to have_http_status(:success)
-    end
-  end
-
-end
